+++ src/main/algorithm/MemMem.h
@@ -149,15 +149,15 @@ public:
     }
 
     /* Searching */
-    int search(const char_type * text, size_type text_len,
-               const char_type * pattern, size_type pattern_len) const {
+    Long search(const char_type * text, size_type text_len,
+                const char_type * pattern, size_type pattern_len) const {
         assert(text != nullptr);
         assert(pattern != nullptr);
         const char_type * haystack = (const char_type *)memmem(
                                         (const void *)text, text_len * sizeof(char_type),
                                         (const void *)pattern, pattern_len * sizeof(char_type));
         if (likely(haystack != nullptr))
-            return (int)(haystack - text);
+            return (Long)(haystack - text);
         else
             return Status::NotFound;
     }
